+++ numbers/randExp_test.go
@@ -1,0 +1,12 @@
+package numbers
+
+import (
+	"testing"
+)
+
+// Uncomment to print random variates for plotting
+func TestRandGeometric(t *testing.T) {
+	for i := 0; i < 10; i++ {
+		//fmt.Println(RandGeometric(0.5))
+	}
+}
